new changes

+++ libc/open.c
@@ -1,6 +1,6 @@
 #include<libc.h>
 
 int open(char *pathname, int flags) {
-    return syscall2(2, (void*)pathname, flags);
+    return (int)syscall2(2, (void*)pathname, flags);
 }
